bm25 with lucene added
